requirements.txt
Django==4.2
djangorestframework
gunicorn
psycopg2